Update import API keys, modify arguments in data objects, and enhance query filters
The changes made in the codes are mainly modifying existing content for improved functionality. API keys for Weaviate connection is updated with the required commas, and arguments within data objects have been refactored for cleaner representation. Furthermore, updates were made in the 3_query.py file to provide more in-depth filters for query iterations. Minor adjustments are also made on the execution counts and data representation in the workshop examples.